LayoutTests/accessibility/aria-required-expected.txt

This tests that aria-required is a usable attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Elements to test: 35

PASS textfield isRequired is false
PASS textfield_required isRequired is true
PASS textfield_required_ariarequired isRequired is true
PASS textfield_ariarequiredfalse isRequired is false
PASS textfield5 isRequired is true
PASS checkbox isRequired is false
PASS checkbox_required isRequired is true
PASS checkbox_required_ariarequired isRequired is true
PASS checkbox_ariarequiredfalse isRequired is false
PASS checkbox_required_ariarequiredfalse isRequired is false
PASS checkbox_requiredfalse_ariarequiredtrue isRequired is true
PASS select isRequired is false
PASS select_required isRequired is true
PASS select_required_ariarequired isRequired is true
PASS select_ariarequiredfalse isRequired is false
PASS select_required_ariarequiredfalse isRequired is false
PASS select_requiredfalse_ariarequiredtrue isRequired is true
PASS textarea isRequired is false
PASS textarea_required isRequired is true
PASS textarea_required_ariarequired isRequired is true
PASS textarea_ariarequiredfalse isRequired is false
PASS textarea_required_ariarequiredfalse isRequired is false
PASS textarea_requiredfalse_ariarequiredtrue isRequired is true
PASS listbox_ariarequiredtrue isRequired is true
PASS listbox_ariarequiredfalse isRequired is false
PASS listbox isRequired is false
PASS radiogroup_ariarequiredtrue isRequired is true
PASS radiogroup_ariarequiredfalse isRequired is false
PASS radiogroup isRequired is false
PASS spinbutton_ariarequiredtrue isRequired is true
PASS spinbutton_ariarequiredfalse isRequired is false
PASS spinbutton isRequired is false
PASS tree_ariarequiredtrue isRequired is true
PASS tree_ariarequiredfalse isRequired is false
PASS tree isRequired is false
PASS successfullyParsed is true

TEST COMPLETE
